requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohttp==3.9.5
    # via
    #   llama-index-core
    #   llama-index-legacy
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via aiohttp
beautifulsoup4==4.12.3
    # via llama-index-readers-file
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   nltk
    #   typer
    #   uvicorn
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.7
    # via
    #   llama-index-core
    #   llama-index-legacy
dateparser==1.2.0
    # via semantic-catalogue-demo (pyproject.toml)
deprecated==1.2.14
    # via
    #   llama-index-core
    #   llama-index-legacy
dirtyjson==1.0.8
    # via
    #   llama-index-core
    #   llama-index-legacy
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via email-validator
email-validator==2.1.2
    # via fastapi
fastapi==0.111.0
    # via semantic-catalogue-demo (pyproject.toml)
fastapi-cli==0.0.4
    # via fastapi
fonttools==4.53.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.0
    # via
    #   llama-index-core
    #   llama-index-legacy
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   llama-index-core
    #   llama-index-legacy
    #   llamaindex-py-client
    #   openai
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.4
    # via fastapi
joblib==1.4.2
    # via nltk
kiwisolver==1.4.5
    # via matplotlib
llama-index==0.10.46
    # via semantic-catalogue-demo (pyproject.toml)
llama-index-agent-openai==0.2.7
    # via
    #   llama-index
    #   llama-index-program-openai
llama-index-cli==0.1.12
    # via llama-index
llama-index-core==0.10.46
    # via
    #   llama-index
    #   llama-index-agent-openai
    #   llama-index-cli
    #   llama-index-embeddings-openai
    #   llama-index-indices-managed-llama-cloud
    #   llama-index-llms-openai
    #   llama-index-multi-modal-llms-openai
    #   llama-index-program-openai
    #   llama-index-question-gen-openai
    #   llama-index-readers-file
    #   llama-index-readers-llama-parse
    #   llama-index-vector-stores-pinecone
    #   llama-parse
llama-index-embeddings-openai==0.1.10
    # via
    #   llama-index
    #   llama-index-cli
llama-index-indices-managed-llama-cloud==0.1.6
    # via llama-index
llama-index-legacy==0.9.48
    # via llama-index
llama-index-llms-openai==0.1.22
    # via
    #   llama-index
    #   llama-index-agent-openai
    #   llama-index-cli
    #   llama-index-multi-modal-llms-openai
    #   llama-index-program-openai
    #   llama-index-question-gen-openai
llama-index-multi-modal-llms-openai==0.1.6
    # via llama-index
llama-index-program-openai==0.1.6
    # via
    #   llama-index
    #   llama-index-question-gen-openai
llama-index-question-gen-openai==0.1.3
    # via llama-index
llama-index-readers-file==0.1.25
    # via
    #   semantic-catalogue-demo (pyproject.toml)
    #   llama-index
llama-index-readers-llama-parse==0.1.4
    # via llama-index
llama-index-vector-stores-pinecone==0.1.7
    # via semantic-catalogue-demo (pyproject.toml)
llama-parse==0.4.4
    # via
    #   semantic-catalogue-demo (pyproject.toml)
    #   llama-index-readers-llama-parse
llamaindex-py-client==0.1.19
    # via
    #   llama-index-core
    #   llama-index-indices-managed-llama-cloud
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
marshmallow==3.21.3
    # via dataclasses-json
matplotlib==3.9.0
    # via seaborn
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via
    #   llama-index-core
    #   llama-index-legacy
networkx==3.3
    # via
    #   llama-index-core
    #   llama-index-legacy
nltk==3.8.1
    # via
    #   llama-index-core
    #   llama-index-legacy
numpy==1.26.4
    # via
    #   contourpy
    #   llama-index-core
    #   llama-index-legacy
    #   matplotlib
    #   pandas
    #   seaborn
openai==1.34.0
    # via
    #   llama-index-agent-openai
    #   llama-index-core
    #   llama-index-legacy
orjson==3.10.5
    # via fastapi
packaging==24.1
    # via
    #   marshmallow
    #   matplotlib
pandas==2.2.2
    # via
    #   llama-index-core
    #   llama-index-legacy
    #   seaborn
pillow==10.3.0
    # via
    #   llama-index-core
    #   matplotlib
pinecone-client==3.2.2
    # via llama-index-vector-stores-pinecone
pydantic==2.7.4
    # via
    #   fastapi
    #   llamaindex-py-client
    #   openai
    #   pydantic-settings
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.3
    # via semantic-catalogue-demo (pyproject.toml)
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
pypdf==4.2.0
    # via llama-index-readers-file
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via
    #   dateparser
    #   pandas
pyyaml==6.0.1
    # via
    #   llama-index-core
    #   uvicorn
regex==2024.5.15
    # via
    #   dateparser
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   semantic-catalogue-demo (pyproject.toml)
    #   llama-index-core
    #   llama-index-legacy
    #   tiktoken
rich==13.7.1
    # via typer
seaborn==0.13.2
    # via semantic-catalogue-demo (pyproject.toml)
setuptools==70.0.0
    # via semantic-catalogue-demo (pyproject.toml)
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.30
    # via
    #   llama-index-core
    #   llama-index-legacy
starlette==0.37.2
    # via fastapi
striprtf==0.0.26
    # via llama-index-readers-file
tenacity==8.3.0
    # via
    #   llama-index-core
    #   llama-index-legacy
tiktoken==0.7.0
    # via
    #   llama-index-core
    #   llama-index-legacy
tqdm==4.66.4
    # via
    #   llama-index-core
    #   nltk
    #   openai
    #   pinecone-client
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   llama-index-core
    #   llama-index-legacy
    #   openai
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
    #   llama-index-legacy
tzdata==2024.1
    # via pandas
tzlocal==5.2
    # via dateparser
ujson==5.10.0
    # via fastapi
urllib3==2.2.2
    # via
    #   pinecone-client
    #   requests
uvicorn==0.30.1
    # via fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
wrapt==1.16.0
    # via
    #   deprecated
    #   llama-index-core
yarl==1.9.4
    # via aiohttp
